requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.5.0
pycocotools>=2.0.6
tqdm>=4.65.0
matplotlib>=3.7.0
nltk>=3.8.1

